Remove jobs_runner_asyncio.py and integrate its functionality into Jobs class
- Removed the separate JobsRunnerAsyncio class
- Integrated runner functionality directly into the Jobs._execute_with_remote_cache method
- Updated tests to work without the removed file
- Simplified code flow and reduced unnecessary indirection
