+++ tests/jobs/test_Jobs.py
@@ -265,20 +265,12 @@ def test_handle_model_exception():
 
 
 def test_jobs_bucket_creator(valid_job):
-    # from edsl.jobs.runners.job_runners_registry import JobsRunnersRegistry
-    # JobRunner = JobsRunnersRegistry["asyncio"](jobs=valid_job)
-    from edsl.jobs.jobs_runner_asyncio import JobsRunnerAsyncio
-
+    # Create a bucket collection directly
     bc_to_use = valid_job.create_bucket_collection()
 
-    from edsl.jobs.data_structures import RunEnvironment
-
-    re = RunEnvironment()
-    re.bucket_collection = bc_to_use
-
-    bc = JobsRunnerAsyncio(jobs=valid_job, environment=re).environment.bucket_collection
-    assert bc[valid_job.models[0]].requests_bucket.tokens > 10
-    assert bc[valid_job.models[0]].tokens_bucket.tokens > 10
+    # Test bucket collection directly without using the removed JobsRunnerAsyncio class
+    assert bc_to_use[valid_job.models[0]].requests_bucket.tokens > 10
+    assert bc_to_use[valid_job.models[0]].tokens_bucket.tokens > 10
